Added entrypoint hooks to extend this image
Most of the changes are inspired by the Postgres Docker image,
for the purpose of allowing this image to be extended like
Postgres, without keeping our own copy of gerrit-start.sh.
This is especially useful with plugins, where we want to
initialise and configure the plugins as part of the first init
call.

Specifically, the following alterations have been made:

 * Dockerfile runs wholly as root, using gosu to execute as
   ${GERRIT_USER}.
 * Added /docker-entrypoint-init.d/ directory and hook, to allow
   further customisation by other images.
 * Added $GERRIT_INIT_ARGS to the end of the gerrit init calls, to
   allow plugin installation to occur as per gerrit documentation.
 * Added documentation regarding how to extend this image.
 * Moved gerrit-entrypoint.sh and gerrit-start.sh to /, so it's harder
   to lose them when playing silly buggers with volume mounting.
 * Added set_gerrit_config and set_secure_config functions to
   gerrit-entrypoint.sh to save having gosu calls everywhere on already
   long lines.