override default start action with $GERRIT_START_ACTION
Gerrit is started with the 'daemon' action which logs error messages
to the console.  The 'supervise' action is very similar to 'daemon'
except that error messages are persisted to $GERRIT_SITE/logs/error_log.
Allow the start action to be overridden with $GERRIT_START_ACTION if the
user has preference other than the default.